+++ pkg/gateway/cache/triggers.go
@@ -50,35 +50,6 @@ func (c *GatewayCache) isRoutableHTTPService(service client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-
-	//for _, r := range c.getResourcesFromCache(informers.HTTPRoutesResourceType, false) {
-	//	r := r.(*gwv1.HTTPRoute)
-	//	for _, rule := range r.Spec.Rules {
-	//		for _, backend := range rule.BackendRefs {
-	//			if c.isRefToService(r, backend.BackendObjectReference, service) {
-	//				return true
-	//			}
-	//
-	//			for _, filter := range backend.Filters {
-	//				if filter.Type == gwv1.HTTPRouteFilterRequestMirror {
-	//					if c.isRefToService(r, filter.RequestMirror.BackendRef, service) {
-	//						return true
-	//					}
-	//				}
-	//			}
-	//		}
-	//
-	//		for _, filter := range rule.Filters {
-	//			if filter.Type == gwv1.HTTPRouteFilterRequestMirror {
-	//				if c.isRefToService(r, filter.RequestMirror.BackendRef, service) {
-	//					return true
-	//				}
-	//			}
-	//		}
-	//	}
-	//}
-
-	//return false
 }
 
 // no need to check ReferenceGrant here
@@ -92,34 +63,6 @@ func (c *GatewayCache) isRoutableGRPCService(service client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-	//for _, r := range c.getResourcesFromCache(informers.GRPCRoutesResourceType, false) {
-	//	r := r.(*gwv1.GRPCRoute)
-	//	for _, rule := range r.Spec.Rules {
-	//		for _, backend := range rule.BackendRefs {
-	//			if c.isRefToService(r, backend.BackendObjectReference, service) {
-	//				return true
-	//			}
-	//
-	//			for _, filter := range backend.Filters {
-	//				if filter.Type == gwv1.GRPCRouteFilterRequestMirror {
-	//					if c.isRefToService(r, filter.RequestMirror.BackendRef, service) {
-	//						return true
-	//					}
-	//				}
-	//			}
-	//		}
-	//
-	//		for _, filter := range rule.Filters {
-	//			if filter.Type == gwv1.GRPCRouteFilterRequestMirror {
-	//				if c.isRefToService(r, filter.RequestMirror.BackendRef, service) {
-	//					return true
-	//				}
-	//			}
-	//		}
-	//	}
-	//}
-	//
-	//return false
 }
 
 // no need to check ReferenceGrant here
@@ -133,18 +76,6 @@ func (c *GatewayCache) isRoutableTLSService(service client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-	//for _, r := range c.getResourcesFromCache(informers.TLSRoutesResourceType, false) {
-	//	r := r.(*gwv1alpha2.TLSRoute)
-	//	for _, rule := range r.Spec.Rules {
-	//		for _, backend := range rule.BackendRefs {
-	//			if c.isRefToService(r, backend.BackendObjectReference, service) {
-	//				return true
-	//			}
-	//		}
-	//	}
-	//}
-	//
-	//return false
 }
 
 // no need to check ReferenceGrant here
@@ -158,18 +89,6 @@ func (c *GatewayCache) isRoutableTCPService(service client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-	//for _, r := range c.getResourcesFromCache(informers.TCPRoutesResourceType, false) {
-	//	r := r.(*gwv1alpha2.TCPRoute)
-	//	for _, rule := range r.Spec.Rules {
-	//		for _, backend := range rule.BackendRefs {
-	//			if c.isRefToService(r, backend.BackendObjectReference, service) {
-	//				return true
-	//			}
-	//		}
-	//	}
-	//}
-	//
-	//return false
 }
 
 // no need to check ReferenceGrant here
@@ -183,18 +102,6 @@ func (c *GatewayCache) isRoutableUDPService(service client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-	//for _, r := range c.getResourcesFromCache(informers.UDPRoutesResourceType, false) {
-	//	r := r.(*gwv1alpha2.UDPRoute)
-	//	for _, rule := range r.Spec.Rules {
-	//		for _, backend := range rule.BackendRefs {
-	//			if c.isRefToService(r, backend.BackendObjectReference, service) {
-	//				return true
-	//			}
-	//		}
-	//	}
-	//}
-	//
-	//return false
 }
 
 // no need to check ReferenceGrant here
@@ -237,18 +144,7 @@ func (c *GatewayCache) isEffectiveTargetRef(policy client.Object, targetRef gwv1
 		}
 
 		return gwutils.IsActiveGateway(gw)
-		//gateways := c.getActiveGateways()
-		//if len(gateways) == 0 {
-		//	return false
-		//}
-		//
-		//for _, gateway := range gateways {
-		//	if gwutils.IsTargetRefToTarget(policy, targetRef, gateway) {
-		//		return true
-		//	}
-		//}
 	case constants.GatewayAPIHTTPRouteKind:
-		//httproutes := c.getResourcesFromCache(informers.HTTPRoutesResourceType, false)
 		route := &gwv1.HTTPRoute{}
 		if err := c.client.Get(context.Background(), key, route); err != nil {
 			log.Error().Msgf("Failed to get HTTPRoute: %v", err)
@@ -256,25 +152,7 @@ func (c *GatewayCache) isEffectiveTargetRef(policy client.Object, targetRef gwv1
 		}
 
 		return true
-		//list := &gwv1.HTTPRouteList{}
-		//if err := c.client.List(context.Background(), list, &client.ListOptions{}); err != nil {
-		//	log.Error().Msgf("Failed to list HTTPRoutes: %v", err)
-		//	return false
-		//}
-		//
-		//httproutes := gwutils.ToSlicePtr(list.Items)
-		//
-		//if len(httproutes) == 0 {
-		//	return false
-		//}
-		//
-		//for _, route := range httproutes {
-		//	if gwutils.IsTargetRefToTarget(policy, targetRef, route) {
-		//		return true
-		//	}
-		//}
 	case constants.GatewayAPIGRPCRouteKind:
-
 		route := &gwv1.GRPCRoute{}
 		if err := c.client.Get(context.Background(), key, route); err != nil {
 			log.Error().Msgf("Failed to get GRPCRoute: %v", err)
@@ -282,24 +160,6 @@ func (c *GatewayCache) isEffectiveTargetRef(policy client.Object, targetRef gwv1
 		}
 
 		return true
-		//grpcroutes := c.getResourcesFromCache(informers.GRPCRoutesResourceType, false)
-		//list := &gwv1.GRPCRouteList{}
-		//if err := c.client.List(context.Background(), list, &client.ListOptions{}); err != nil {
-		//	log.Error().Msgf("Failed to list GRPCRoutes: %v", err)
-		//	return false
-		//}
-		//
-		//grpcroutes := gwutils.ToSlicePtr(list.Items)
-		//
-		//if len(grpcroutes) == 0 {
-		//	return false
-		//}
-		//
-		//for _, route := range grpcroutes {
-		//	if gwutils.IsTargetRefToTarget(policy, targetRef, route) {
-		//		return true
-		//	}
-		//}
 	}
 
 	return false
@@ -320,8 +180,6 @@ func (c *GatewayCache) isRoutableTargetService(owner client.Object, targetRef gw
 
 // no need to check ReferenceGrant here
 func (c *GatewayCache) isSecretReferred(secret client.ObjectKey) bool {
-	//ctx := context.TODO()
-
 	list := &gwv1.GatewayList{}
 	if err := c.client.List(context.Background(), list, &client.ListOptions{
 		FieldSelector: fields.OneTermEqualSelector(constants.SecretGatewayIndex, secret.String()),
@@ -334,42 +192,6 @@ func (c *GatewayCache) isSecretReferred(secret client.ObjectKey) bool {
 		return true
 	}
 
-	//for _, gw := range c.getActiveGateways() {
-	//	for _, l := range gw.Spec.Listeners {
-	//		switch l.Protocol {
-	//		case gwv1.HTTPSProtocolType, gwv1.TLSProtocolType:
-	//			if l.TLS == nil {
-	//				continue
-	//			}
-	//
-	//			if l.TLS.Mode == nil || *l.TLS.Mode == gwv1.TLSModeTerminate {
-	//				if len(l.TLS.CertificateRefs) > 0 {
-	//					for _, ref := range l.TLS.CertificateRefs {
-	//						if c.isRefToSecret(gw, ref, secret) {
-	//							return true
-	//						}
-	//					}
-	//				}
-	//
-	//				if l.TLS.FrontendValidation != nil && len(l.TLS.FrontendValidation.CACertificateRefs) > 0 {
-	//					for _, ref := range l.TLS.FrontendValidation.CACertificateRefs {
-	//						ref := gwv1.SecretObjectReference{
-	//							Group:     ptr.To(ref.Group),
-	//							Kind:      ptr.To(ref.Kind),
-	//							Name:      ref.Name,
-	//							Namespace: ref.Namespace,
-	//						}
-	//
-	//						if c.isRefToSecret(gw, ref, secret) {
-	//							return true
-	//						}
-	//					}
-	//				}
-	//			}
-	//		}
-	//	}
-	//}
-
 	policies := &gwpav1alpha1.UpstreamTLSPolicyList{}
 	if err := c.client.List(context.Background(), policies, &client.ListOptions{
 		FieldSelector: fields.OneTermEqualSelector(constants.SecretUpstreamTLSPolicyIndex, secret.String()),
@@ -379,30 +201,6 @@ func (c *GatewayCache) isSecretReferred(secret client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-
-	//for _, ut := range c.getResourcesFromCache(informers.UpstreamTLSPoliciesResourceType, false) {
-	//	ut := ut.(*gwpav1alpha1.UpstreamTLSPolicy)
-	//
-	//	if ut.Spec.DefaultConfig != nil {
-	//		if c.isRefToSecret(ut, ut.Spec.DefaultConfig.CertificateRef, secret) {
-	//			return true
-	//		}
-	//	}
-	//
-	//	if len(ut.Spec.Ports) > 0 {
-	//		for _, port := range ut.Spec.Ports {
-	//			if port.Config == nil {
-	//				continue
-	//			}
-	//
-	//			if c.isRefToSecret(ut, port.Config.CertificateRef, secret) {
-	//				return true
-	//			}
-	//		}
-	//	}
-	//}
-	//
-	//return false
 }
 
 // no need to check ReferenceGrant here
@@ -417,33 +215,4 @@ func (c *GatewayCache) isConfigMapReferred(cm client.ObjectKey) bool {
 	}
 
 	return len(list.Items) > 0
-
-	//for _, gw := range c.getActiveGateways() {
-	//	for _, l := range gw.Spec.Listeners {
-	//		switch l.Protocol {
-	//		case gwv1.HTTPSProtocolType, gwv1.TLSProtocolType:
-	//			if l.TLS == nil {
-	//				continue
-	//			}
-	//
-	//			if l.TLS.Mode == nil || *l.TLS.Mode == gwv1.TLSModeTerminate {
-	//				if l.TLS.FrontendValidation == nil {
-	//					continue
-	//				}
-	//
-	//				if len(l.TLS.FrontendValidation.CACertificateRefs) == 0 {
-	//					continue
-	//				}
-	//
-	//				for _, ref := range l.TLS.FrontendValidation.CACertificateRefs {
-	//					if c.isRefToConfigMap(gw, ref, cm) {
-	//						return true
-	//					}
-	//				}
-	//			}
-	//		}
-	//	}
-	//}
-	//
-	//return false
 }
