+++ pkg/gateway/cache/types.go
@@ -47,7 +47,6 @@ type Cache interface {
 
 type serviceContext struct {
 	svcPortName fgw.ServicePortName
-	//filters     []routecfg.Filter
 }
 
 type endpointContext struct {
@@ -56,24 +55,6 @@ type endpointContext struct {
 }
 
 type calculateEndpointsFunc func(svc *corev1.Service, port *int32) map[string]fgw.Endpoint
-
-//type globalPolicyAttachments struct {
-//	rateLimits      map[gwpkg.PolicyMatchType][]*gwpav1alpha1.RateLimitPolicy
-//	accessControls  map[gwpkg.PolicyMatchType][]*gwpav1alpha1.AccessControlPolicy
-//	faultInjections map[gwpkg.PolicyMatchType][]*gwpav1alpha1.FaultInjectionPolicy
-//}
-//
-//type routePolicies struct {
-//	hostnamesRateLimits      []gwpav1alpha1.RateLimitPolicy
-//	httpRouteRateLimits      []gwpav1alpha1.RateLimitPolicy
-//	grpcRouteRateLimits      []gwpav1alpha1.RateLimitPolicy
-//	hostnamesAccessControls  []gwpav1alpha1.AccessControlPolicy
-//	httpRouteAccessControls  []gwpav1alpha1.AccessControlPolicy
-//	grpcRouteAccessControls  []gwpav1alpha1.AccessControlPolicy
-//	hostnamesFaultInjections []gwpav1alpha1.FaultInjectionPolicy
-//	httpRouteFaultInjections []gwpav1alpha1.FaultInjectionPolicy
-//	grpcRouteFaultInjections []gwpav1alpha1.FaultInjectionPolicy
-//}
 
 var (
 	log = logger.New("fsm-gateway/cache")
