+++ pkg/gateway/policy/status/policy_status_processor.go
@@ -66,11 +66,6 @@ func (p *PolicyStatusProcessor) Process(ctx context.Context, policy client.Objec
 		return InvalidCondition(policy, fmt.Sprintf("Invalid target reference kind, only %q are supported", strings.Join(p.supportedKinds(), ",")))
 	}
 
-	//referenceGrants := p.Informer.GetGatewayResourcesFromCache(informers.ReferenceGrantResourceType, false)
-	//if !gwutils.HasAccessToTargetRef(policy, targetRef, referenceGrants) {
-	//	return NoAccessCondition(policy, fmt.Sprintf("Cross namespace reference to target %s/%s/%s is not allowed", targetRef.Kind, ns(targetRef.Namespace), targetRef.Name))
-	//}
-
 	key := types.NamespacedName{
 		Namespace: gwutils.Namespace(targetRef.Namespace, policy.GetNamespace()),
 		Name:      string(targetRef.Name),
@@ -198,11 +193,3 @@ func (p *PolicyStatusProcessor) supportedKinds() []string {
 
 	return kinds
 }
-
-//func ns(namespace *gwv1.Namespace) string {
-//	if namespace == nil {
-//		return ""
-//	}
-//
-//	return string(*namespace)
-//}
