+++ pkg/gateway/utils/services.go
@@ -6,6 +6,7 @@ import (
 	"github.com/flomesh-io/fsm/pkg/constants"
 )
 
+// IsValidTargetRefToGroupKindOfService checks if the target reference is valid to the group kind of service
 func IsValidTargetRefToGroupKindOfService(ref gwv1alpha2.NamespacedPolicyTargetReference) bool {
 	if (ref.Kind == constants.KubernetesServiceKind && ref.Group == constants.KubernetesCoreGroup) ||
 		(ref.Kind == constants.FlomeshAPIServiceImportKind && ref.Group == constants.FlomeshMCSAPIGroup) {
