+++ pkg/ingress/providers/pipy/cache/service.go
@@ -249,25 +249,6 @@ func (t *ServiceChangeTracker) shouldSkipService(svc *corev1.Service) bool {
 	return false
 }
 
-//func (sct *ServiceChangeTracker) serviceImportExists(svc *corev1.Service) bool {
-//	_, err := sct.informers.GetListers().ServiceImport.
-//		ServiceImports(svc.Namespace).
-//		Get(svc.Name)
-//	if err != nil {
-//		if errors.IsNotFound(err) {
-//			// do nothing, not exists, go ahead and check svc
-//			log.Info().Msgf("ServiceImport %s/%s doesn't exist", svc.Namespace, svc.Name)
-//			return false
-//		}
-//
-//		log.Warn().Msgf("Failed to get ServiceImport %s/%s, %s", svc.Namespace, svc.Name, err)
-//
-//		return false
-//	}
-//
-//	return true
-//}
-
 func (sm *ServiceMap) apply(changes *ServiceChangeTracker) {
 	changes.lock.Lock()
 	defer changes.lock.Unlock()
